requirements.txt
Jinja2==3.1.2
jmespath==1.0.1
jsonpath==0.82
pytest==7.2.0
PyYAML==6.0
requests==2.28.1